Refactor editor i/o. Tabs now show filenames for locally opened files
